requirements.txt
click # CLI tool
python-novaclient # OpenStack
pycrypto # SSH RSA KEY
